Trigger only time complications update for TIME_VALUE
