+++ wear/src/main/java/de/michelinside/glucodatahandler/complications/ShortTextComplication.kt
@@ -3,7 +3,6 @@ package de.michelinside.glucodatahandler
 import androidx.wear.watchface.complications.data.*
 import de.michelinside.glucodatahandler.common.ReceiveData
 import de.michelinside.glucodatahandler.common.Utils
-import java.util.*
 
 open class ShortClucoseComplication:  BgValueComplicationService() {
 }
@@ -34,39 +33,6 @@ class ShortGlucoseWithDeltaAndTrendComplication: ShortClucoseComplication() {
         )
             .setSmallImage(glucoseImage())
             .setTapAction(getTapAction())
-            .build()
-    }
-}
-
-class LongGlucoseWithDeltaAndTrendAndTimeComplication: ShortClucoseComplication() {
-    override fun getLongTextComplicationData(): ComplicationData {
-        return LongTextComplicationData.Builder(
-            plainText(" Δ: " + ReceiveData.getDeltaAsString()),
-            descriptionText()
-        )
-            .setTitle(timeText())
-            .setSmallImage(getGlucoseTrendImage())
-            .setTapAction(getTapAction())
-            .build()
-    }
-}
-
-class TimeStampComplication: BgValueComplicationService() {
-    override fun getText(): PlainComplicationText = timeText(true)
-    override fun getIcon(): MonochromaticImage = glucoseIcon()
-
-    override fun getRangeValueComplicationData(): ComplicationData {
-        val time = Date(ReceiveData.time)
-
-        val value = if (ReceiveData.isObsolete(3600)) -1F else (time.minutes * 60 + time.seconds).toFloat()
-        return RangedValueComplicationData.Builder(
-            value = value,
-            min = 0F,
-            max = 3600F,
-            contentDescription = descriptionText()
-        )
-            .setText(getText())
-            .setMonochromaticImage(getIcon())
             .build()
     }
 }
@@ -106,10 +72,6 @@ class ShortDeltaWithTrendComplication: ShortDeltaComplication() {
 
 class ShortDeltaWithIconComplication: ShortDeltaComplication() {
     override fun getIcon(): MonochromaticImage = deltaIcon()
-}
-
-class ShortDeltaWithTimeComplication: ShortDeltaComplication() {
-    override fun getTitle(): PlainComplicationText = timeText(true)
 }
 
 open class ShortTrendComplication: ShortClucoseComplication() {
